fix(core): Remove and update addDataPointEventListener
Remove eventlistener from base-axis-chart reimplement in bar-chart

Enchancement #97: Remove addDataPointEventListener() from base-axis-chart.ts
