+++ packages/core/src/bar-chart.ts
@@ -1,5 +1,5 @@
 // D3 Imports
-import { select } from "d3-selection";
+import { select, mouse } from "d3-selection";
 import { scaleBand } from "d3-scale";
 import { min } from "d3-array";
 
@@ -244,5 +244,40 @@ export class BarChart extends BaseAxisChart {
 		this.updateElements(false, null, g);
 
 		super.resizeChart();
+	}
+
+	addDataPointEventListener() {
+		const self = this;
+		const { accessibility } = this.options;
+
+		this.svg.selectAll("rect")
+			.on("click", function(d) {
+				self.dispatchEvent("bar-onClick", d);
+			})
+			.on("mouseover", function(d) {
+				select(this)
+					.attr("stroke-width", Configuration.bars.mouseover.strokeWidth)
+					.attr("stroke", self.colorScale[d.datasetLabel](d.label))
+					.attr("stroke-opacity", Configuration.bars.mouseover.strokeOpacity);
+
+				self.showTooltip(d, this);
+				self.reduceOpacity(this);
+			})
+			.on("mousemove", function(d) {
+				const tooltipRef = select(self.holder).select("div.chart-tooltip");
+
+				const relativeMousePosition = mouse(self.holder as HTMLElement);
+				tooltipRef.style("left", relativeMousePosition[0] + Configuration.tooltip.magicLeft2 + "px")
+					.style("top", relativeMousePosition[1] + "px");
+			})
+			.on("mouseout", function(d) {
+				const { strokeWidth, strokeWidthAccessible } = Configuration.bars.mouseout;
+				select(this)
+					.attr("stroke-width", accessibility ? strokeWidthAccessible : strokeWidth)
+					.attr("stroke", accessibility ? self.colorScale[d.datasetLabel](d.label) : "none")
+					.attr("stroke-opacity", Configuration.bars.mouseout.strokeOpacity);
+
+				self.hideTooltip();
+			});
 	}
 }
